+++ src/cli/index.ts
@@ -1,4 +1,5 @@
 export { appCheckCommandDefinition } from './google-app-check.js';
 export { firebaseStorageCommandDefinition } from './google-firebase-storage.js';
 export { firestoreCommandDefinition } from './google-firestore.js';
+export { identityPlatformCommandDefinition } from './google-identity-platform.js';
 export { googleCommandDefinition } from './google.js';