+++ src/functions/index.ts
@@ -12,6 +12,8 @@ import { EmulatorStopForSpanner } from './emulator-stop-spanner.js';
 import { GoogleAppCheckGenerateToken } from './google-app-check-generate-token.js';
 import { GoogleFirebaseStorageMergeRules } from './google-firebase-storage-merge-rules.js';
 import { GoogleFirestoreMergeRules } from './google-firestore-merge-rules.js';
+import { GoogleIdentityPlatformGenerateCustomToken } from './google-identity-platform-generate-custom-token.js';
+import { GoogleIdentityPlatformGenerateToken } from './google-identity-platform-generate-token.js';
 import { GoogleServicesEnable } from './google-services-enable.js';
 import { GoogleSpannerListDatabases } from './google-spanner-list-databases.js';
 import { ProjectGetArtefactDestinationForServiceContainer } from './project-get-artefact-destination-service-container.js';
@@ -32,6 +34,8 @@ export function registerFunctions(context: ModuleRegistrationContext) {
     GoogleAppCheckGenerateToken,
     GoogleFirebaseStorageMergeRules,
     GoogleFirestoreMergeRules,
+    GoogleIdentityPlatformGenerateCustomToken,
+    GoogleIdentityPlatformGenerateToken,
     GoogleServicesEnable,
     GoogleSpannerListDatabases,
     ProjectGetArtefactDestinationForServiceContainer,